started to work on the regression network
